Remove blocking from Browser
