+++ packages/browser/tests/unit/go.spec.ts
@@ -148,58 +148,4 @@ describe("go", () => {
       });
     });
   });
-
-  describe("with user confirmation", () => {
-    it("calls response handler after the user confirms the navigation", done => {
-      const testHistory = Browser();
-      let setup = false;
-      const router = ignoreFirstCall(function(pending) {
-        pending.finish();
-        if (!setup) {
-          return;
-        }
-        expect(testHistory.location).toMatchObject({
-          pathname: "/one",
-          key: "0.0"
-        });
-        done();
-      });
-      const confirm = (info, confirm, prevent) => {
-        confirm();
-      };
-      testHistory.confirmWith(confirm);
-      testHistory.respondWith(router);
-
-      testHistory.navigate("/two", "push"); // 1.0
-      testHistory.navigate("/three", "push"); // 2.0
-      setup = true;
-      testHistory.go(-2);
-    });
-
-    it("does not call response handler when the user prevents the navigation", done => {
-      const testHistory = Browser();
-      const confirm = (info, confirm, prevent) => {
-        prevent();
-      };
-      function router(pending) {
-        pending.finish();
-      }
-      testHistory.respondWith(router);
-
-      testHistory.navigate("/two", "push"); // 1.0
-      testHistory.navigate("/three", "push"); // 2.0
-      // don't add function until we have setup the history
-      testHistory.confirmWith(confirm);
-      testHistory.go(-2);
-
-      // need to wait for window.history.go to emit a popstate event
-      setTimeout(() => {
-        expect(testHistory.location).toMatchObject({
-          pathname: "/three",
-          key: "2.0"
-        });
-        done();
-      }, 50);
-    });
-  });
 });
