+++ packages/browser/types/Browser.d.ts
@@ -1,0 +1,3 @@
+import { BrowserHistory, Options } from "./types";
+declare function Browser<Q = string>(options?: Options<Q>): BrowserHistory<Q>;
+export { Browser };
