+++ packages/browser/types/types.d.ts
@@ -1,0 +1,4 @@
+import { History, LocationComponents, PartialLocation, SessionLocation, AnyLocation, Location, LocationUtilOptions } from "@hickory/root";
+export { History, SessionLocation, PartialLocation, AnyLocation, Location, LocationComponents };
+export declare type Options<Q> = LocationUtilOptions<Q>;
+export declare type BrowserHistory<Q> = History<Q>;
